+++ src/private/sources/schema/reader.rs
@@ -3,7 +3,7 @@
 use crate::private::sources::retry::BackoffStrategy;
 use crate::private::sources::schema::error::SchemaException;
 use crate::private::sources::Read;
-use crate::private::types::policy_store_id::PolicyStoreId;
+use crate::private::types::policy_selector::PolicySelector;
 use async_trait::async_trait;
 use aws_sdk_verifiedpermissions::operation::get_schema::{GetSchemaError, GetSchemaOutput};
 use aws_sdk_verifiedpermissions::Client;
@@ -52,13 +52,13 @@ impl GetSchema {
 
 #[async_trait]
 impl Read for GetSchema {
-    type Input = PolicyStoreId;
+    type Input = PolicySelector;
     type Output = GetSchemaOutput;
     type Exception = SchemaException;
 
     #[instrument(skip(self), err(Debug))]
-    async fn read(&self, policy_store_id: Self::Input) -> Result<Self::Output, Self::Exception> {
-        Ok(self.get_schema(&policy_store_id.to_string()).await?)
+    async fn read(&self, policy_selector: Self::Input) -> Result<Self::Output, Self::Exception> {
+        Ok(self.get_schema(&policy_selector.id().to_string()).await?)
     }
 }
 
@@ -68,7 +68,7 @@ mod test {
     use crate::private::sources::schema::reader::GetSchema;
     use crate::private::sources::test::{build_client, build_empty_event, build_event, StatusCode};
     use crate::private::sources::Read;
-    use crate::private::types::policy_store_id::PolicyStoreId;
+    use crate::private::types::policy_selector::PolicySelector;
     use chrono::Utc;
     use serde::{Deserialize, Serialize};
 
@@ -91,41 +91,41 @@ mod test {
 
     #[tokio::test]
     async fn get_schema_200() {
-        let policy_store_id = PolicyStoreId::from("ps-1".to_string());
+        let policy_selector = PolicySelector::from("ps-1".to_string());
         let schema = "some schema";
 
         let request = GetSchemaRequest {
-            policy_store_id: policy_store_id.to_string(),
+            policy_store_id: policy_selector.id().to_string(),
         };
 
         let response = GetSchemaResponse {
             created_date: Utc::now().to_rfc3339(),
             last_updated_date: Utc::now().to_rfc3339(),
-            policy_store_id: policy_store_id.to_string(),
+            policy_store_id: policy_selector.id().to_string(),
             schema: schema.to_string(),
         };
 
         let events = vec![build_event(&request, &response, StatusCode::OK)];
         let client = build_client(events);
         let schema_reader = GetSchema::new(client, BackoffStrategy::default());
-        let result = schema_reader.read(policy_store_id).await.unwrap();
+        let result = schema_reader.read(policy_selector).await.unwrap();
 
         assert_eq!(response.schema, result.schema);
     }
 
     #[tokio::test]
     async fn get_schema_400() {
-        let policy_store_id = PolicyStoreId::from("ps-1".to_string());
+        let policy_selector = PolicySelector::from("ps-1".to_string());
 
         let request = GetSchemaRequest {
-            policy_store_id: policy_store_id.to_string(),
+            policy_store_id: policy_selector.id().to_string(),
         };
 
         let events = vec![build_empty_event(&request, StatusCode::BAD_REQUEST)];
 
         let client = build_client(events);
         let schema_reader = GetSchema::new(client, BackoffStrategy::default());
-        let result = schema_reader.read(policy_store_id).await;
+        let result = schema_reader.read(policy_selector).await;
         assert!(result.is_err());
     }
 }