+++ src/private/types/mod.rs
@@ -2,6 +2,6 @@
 pub mod aliases;
 pub mod cli_shorthand;
 pub mod policy_id;
+pub mod policy_selector;
 pub mod policy_store_filter;
-pub mod policy_store_id;
 pub mod template_id;