+++ src/private/types/policy_selector.rs
@@ -1,4 +1,4 @@
-//! Structure to represent the `PolicyStoreId` used throughout the crate.
+//! Structure to represent the `PolicySelector` used throughout the crate.
 
 use std::fmt;
 
@@ -6,13 +6,13 @@ use crate::public::policy_set_provider::ProviderError;
 
 use super::policy_store_filter::PolicyStoreFilter;
 
-/// This Object wraps the aws verified permissions `PolicyStoreID` which is an unique identifier
+/// This Object wraps the aws verified permissions `PolicySelector` which is an unique identifier
 /// for the policy store.
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct PolicyStoreId(pub String, pub Option<PolicyStoreFilter>);
+pub struct PolicySelector(String, Option<PolicyStoreFilter>);
 
-/// Formats the `PolicyStoreId` using the given formatter.
-impl fmt::Display for PolicyStoreId {
+/// Formats the `PolicySelector` using the given formatter.
+impl fmt::Display for PolicySelector {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         self.0.fmt(f)?;
         if let Some(filter) = &self.1 {
@@ -23,14 +23,14 @@ impl fmt::Display for PolicyStoreId {
     }
 }
 
-/// Allows for conversion from `String` to `PolicyStoreId`
-impl From<String> for PolicyStoreId {
+/// Allows for conversion from `String` to `PolicySelector`
+impl From<String> for PolicySelector {
     fn from(item: String) -> Self {
         Self(item, None)
     }
 }
 
-impl PolicyStoreId {
+impl PolicySelector {
     #[allow(dead_code)]
     pub fn with_cli_filters<T: AsRef<str>>(mut self, filters: T) -> Result<Self, ProviderError> {
         if self.1.is_some() {
@@ -75,51 +75,51 @@ impl PolicyStoreId {
 
 #[cfg(test)]
 mod tests {
-    use crate::private::types::policy_store_id::PolicyStoreId;
+    use crate::private::types::policy_selector::PolicySelector;
     use std::collections::HashMap;
 
     #[test]
     fn policy_store_id_formats_as_expected() {
-        let id = PolicyStoreId::from("id".to_string());
+        let id = PolicySelector::from("id".to_string());
         assert_eq!(id.to_string(), "id");
     }
 
     #[test]
     fn policy_store_id_empty_string() {
-        let id = PolicyStoreId::from(String::new());
+        let id = PolicySelector::from(String::new());
         assert_eq!(id.to_string(), "");
     }
 
     #[test]
     fn policy_store_id_can_be_inserted_into_hashmap() {
-        let mut map: HashMap<PolicyStoreId, i32> = HashMap::new();
-        assert_eq!(map.insert(PolicyStoreId::from("id".to_string()), 1), None);
-        assert_eq!(map.get(&PolicyStoreId::from("id".to_string())), Some(&1));
+        let mut map: HashMap<PolicySelector, i32> = HashMap::new();
+        assert_eq!(map.insert(PolicySelector::from("id".to_string()), 1), None);
+        assert_eq!(map.get(&PolicySelector::from("id".to_string())), Some(&1));
     }
 
     #[test]
     fn policy_store_id_is_cloneable() {
-        let id = PolicyStoreId::from("id".to_string());
+        let id = PolicySelector::from("id".to_string());
         assert_eq!(id.clone(), id);
     }
 
     #[test]
     fn policy_store_id_is_equal_to_another_id_with_same_value() {
-        assert!(PolicyStoreId::from("id".to_string()).eq(&PolicyStoreId::from("id".to_string())));
+        assert!(PolicySelector::from("id".to_string()).eq(&PolicySelector::from("id".to_string())));
     }
 
     #[test]
     fn policy_store_id_is_not_equal_to_another_id_with_different_value() {
         assert!(
-            !PolicyStoreId::from("id".to_string()).eq(&PolicyStoreId::from("other".to_string()))
+            !PolicySelector::from("id".to_string()).eq(&PolicySelector::from("other".to_string()))
         );
     }
 
     #[test]
     fn from_string_to_policy_store_id() {
         assert_eq!(
-            PolicyStoreId::from("ps-1".to_string()),
-            PolicyStoreId::from("ps-1".to_string())
+            PolicySelector::from("ps-1".to_string()),
+            PolicySelector::from("ps-1".to_string())
         );
     }
 
@@ -127,7 +127,7 @@ mod tests {
 
     #[test]
     fn policy_store_id_with_filters_formats_as_expected() {
-        let id = PolicyStoreId::from("id".to_string())
+        let id = PolicySelector::from("id".to_string())
             .with_cli_filters("policyTemplateId=mockPolicyTemplate")
             .expect("CLI filter string should parse correctly");
         assert_eq!(
@@ -138,8 +138,8 @@ mod tests {
 
     #[test]
     fn policy_store_id_with_filters_can_be_inserted_into_hashmap() {
-        let mut map: HashMap<PolicyStoreId, i32> = HashMap::new();
-        let id = PolicyStoreId::from("id".to_string())
+        let mut map: HashMap<PolicySelector, i32> = HashMap::new();
+        let id = PolicySelector::from("id".to_string())
             .with_cli_filters("policyTemplateId=mockPolicyTemplate")
             .expect("CLI filter string should parse correctly");
         let p2 = id.clone();
@@ -149,7 +149,7 @@ mod tests {
 
     #[test]
     fn policy_store_id_with_filters_is_cloneable() {
-        let id = PolicyStoreId::from("id".to_string())
+        let id = PolicySelector::from("id".to_string())
             .with_cli_filters("policyTemplateId=mockPolicyTemplate")
             .expect("CLI filter string should parse correctly");
         assert_eq!(id.clone(), id);
@@ -157,10 +157,10 @@ mod tests {
 
     #[test]
     fn policy_store_id_with_filters_is_equal_to_another_id_with_same_value() {
-        let id = PolicyStoreId::from("id".to_string())
+        let id = PolicySelector::from("id".to_string())
             .with_cli_filters("policyTemplateId=mockPolicyTemplate")
             .expect("CLI filter string should parse correctly");
-        let id2 = PolicyStoreId::from("id".to_string())
+        let id2 = PolicySelector::from("id".to_string())
             .with_cli_filters("policyTemplateId=mockPolicyTemplate")
             .expect("CLI filter string should parse correctly");
         assert!(id.eq(&id2));
@@ -169,7 +169,7 @@ mod tests {
     #[test]
     fn policy_store_id_with_filters_is_not_equal_to_another_id_with_different_value() {
         assert!(
-            !PolicyStoreId::from("id".to_string()).eq(&PolicyStoreId::from("other".to_string()))
+            !PolicySelector::from("id".to_string()).eq(&PolicySelector::from("other".to_string()))
         );
     }
 }