+++ src/public/policy_set_provider.rs
@@ -22,8 +22,8 @@ use crate::private::sources::policy::error::PolicySourceException;
 use crate::private::sources::template::core::{TemplateSource, VerifiedPermissionsTemplateSource};
 use crate::private::sources::template::error::TemplateSourceException;
 use crate::private::translator::avp_to_cedar::Policy;
+use crate::private::types::policy_selector::PolicySelector;
 use crate::private::types::policy_store_filter::{PolicyFilterInputError, PolicyStoreFilter};
-use crate::private::types::policy_store_id::PolicyStoreId;
 
 use super::policy_set_filter::PolicySetFilter;
 
@@ -75,14 +75,14 @@ struct Config {
     /// Gathers templates from Amazon Verified Permissions
     pub template_source: VerifiedPermissionsTemplateSource,
     /// Policy Store Id to gather policies and templates from
-    pub policy_store_id: PolicyStoreId,
+    pub policy_selector: PolicySelector,
 }
 
 /// `PolicySetProvider` structure implements the `SimplePolicySetProvider` trait.
 #[derive(Debug)]
 pub struct PolicySetProvider {
     /// Entities path, stored to allow refreshing from disk.
-    policy_store_id: PolicyStoreId,
+    policy_selector: PolicySelector,
     /// Policy Source
     policy_source: Arc<Mutex<VerifiedPermissionsPolicySource>>,
     /// Policy Source
@@ -100,8 +100,8 @@ impl PolicySetProvider {
     ) -> Result<Self, ProviderError> {
         Self::new(
             ConfigBuilder::default()
-                .policy_store_id(
-                    PolicyStoreId::from(policy_store_id).with_filters(policy_store_filters),
+                .policy_selector(
+                    PolicySelector::from(policy_store_id).with_filters(policy_store_filters),
                 )
                 .policy_source(VerifiedPermissionsPolicySource::from(
                     verified_permissions_client.clone(),
@@ -147,7 +147,7 @@ impl PolicySetProvider {
     #[instrument(skip(config), err(Debug))]
     fn new(config: Config) -> Result<Self, ProviderError> {
         let Config {
-            policy_store_id,
+            policy_selector,
             template_source,
             policy_source,
         } = config;
@@ -156,26 +156,26 @@ impl PolicySetProvider {
         let policy_source = Arc::new(Mutex::new(policy_source));
 
         let mut policy_set = PolicySet::new();
-        let policy_store_id_clone = policy_store_id.clone();
+        let policy_selector_clone = policy_selector.clone();
         let template_source_ref = template_source.clone();
         let templates = task::block_in_place(move || {
             Handle::current().block_on(async move {
                 template_source_ref
                     .lock()
                     .await
-                    .fetch(policy_store_id_clone)
+                    .fetch(policy_selector_clone)
                     .await
             })
         })?;
 
-        let policy_store_id_clone = policy_store_id.clone();
+        let policy_selector_clone = policy_selector.clone();
         let policy_source_ref = policy_source.clone();
         let policies = task::block_in_place(move || {
             Handle::current().block_on(async move {
                 policy_source_ref
                     .lock()
                     .await
-                    .fetch(policy_store_id_clone.clone())
+                    .fetch(policy_selector_clone.clone())
                     .await
             })
         })?;
@@ -222,7 +222,7 @@ impl PolicySetProvider {
         }
 
         Ok(Self {
-            policy_store_id,
+            policy_selector,
             template_source,
             policy_source,
             policy_set: RwLock::new(Arc::new(policy_set)),
@@ -248,7 +248,7 @@ impl UpdateProviderData for PolicySetProvider {
                 .template_source
                 .lock()
                 .await
-                .fetch(self.policy_store_id.clone())
+                .fetch(self.policy_selector.clone())
                 .await
                 .map_err(|e| UpdateProviderDataError::General(Box::new(ProviderError::from(e))))?;
         };
@@ -259,7 +259,7 @@ impl UpdateProviderData for PolicySetProvider {
                 .policy_source
                 .lock()
                 .await
-                .fetch(self.policy_store_id.clone())
+                .fetch(self.policy_selector.clone())
                 .await
                 .map_err(|e| UpdateProviderDataError::General(Box::new(ProviderError::from(e))))?;
         }