+++ CHANGELOG.md
@@ -10,10 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Update to Cedar 4 (requires breaking change)
-- Added `PolicyStoreFilter` as an optional part of the `PolicyStoreId` which then allows a policy store policy
+- Renamed `PolicyStoreId` to `PolicySelector` (because that's what it is now)
+- Added `PolicyStoreFilter` as an optional part of the `PolicySelector` which then allows a policy store policy
   cache to contain only a subset of the policies actually managed by the AVP Policy Store (for example,
   only policies associated with a certain policy template). Note that if the set of policies that one 
-  `PolicyStoreId` identifies intersects with the set of policies that another `PolicyStoreId` identifies,
+  `PolicySelector` identifies intersects with the set of policies that another `PolicySelector` identifies,
   the policies in the intersection will be cached _twice_.
 
 ### Changed